refactor: fix exitAfterDefer in main, drop redundant conversions
- cmd/capsuled/main.go: move daemon body into run() so defer stop()
  unwinds before process exit, instead of os.Exit skipping it.
- Drop 4 redundant byte()/uint64() conversions flagged by golangci-lint's
  unconvert (archivestate_test.go x2, fixed_store_v2_backup.go x2 — the
  second inode comparison wasn't flagged by the linter but has the same
  redundant conversion as the one that was).

No behavior change. go build/vet clean; cmd, archivestate, and
registrationstate package tests pass.

+++ internal/execution/registrationstate/fixed_store_v2_backup.go
@@ -869,7 +869,7 @@ func (store *FixedFileStoreV2) InventoryArchiveArtifacts(
 			stat := info.Sys().(*syscall.Stat_t)
 			report.KnownUnreferenced++
 			candidates = append(candidates, KnownUnreferencedOrphan{
-				digest: segment.Segment.Digest(), device: fmt.Sprint(stat.Dev), inode: uint64(stat.Ino),
+				digest: segment.Segment.Digest(), device: fmt.Sprint(stat.Dev), inode: stat.Ino,
 				checkpoint: loaded.Active.View().CurrentCheckpoint, session: session,
 			})
 		case ArchiveArtifactCrossInstallation:
@@ -971,7 +971,7 @@ func (store *FixedFileStoreV2) DeleteKnownUnreferencedOrphan(
 		return ErrOrphanEvidencePreserved
 	}
 	stat, ok := info.Sys().(*syscall.Stat_t)
-	if !ok || fmt.Sprint(stat.Dev) != candidate.device || uint64(stat.Ino) != candidate.inode {
+	if !ok || fmt.Sprint(stat.Dev) != candidate.device || stat.Ino != candidate.inode {
 		return ErrOrphanEvidencePreserved
 	}
 	if err := failOrphan(faults, FaultOrphanBeforeRemove, false); err != nil {
